requirements.txt
flask
pandas
scikit-learn==1.3.2
tld
googlesearch-python
gunicorn
xgboost